energy-assistant/requirements.txt
streamlit
pypdf
pandas
sentence-transformers
faiss-cpu
requests
transformers
peft
datasets
tiktoken
python 
sentencepiece
tiktoken
